Fix: 카드 추천 서비스 타입 통일

+++ core-api/src/main/java/com/pocketstock/core/portfolio/card/dto/CardRecommendationItem.java
@@ -1,11 +1,12 @@
 package com.pocketstock.core.portfolio.card.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CardRecommendationItem(
         String cardName,
         String cardCompany,
-        Integer annualFee,
+        BigDecimal annualFee,
         List<String> benefits,
         int matchRate
 ) {}
