+++ core-api/src/main/java/com/pocketstock/core/portfolio/card/dto/CardRow.java
@@ -2,11 +2,13 @@ package com.pocketstock.core.portfolio.card.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CardRow {
     private Long id;
     private String cardName;
     private String cardType;
     private String provider;
-    private Integer annualFeeDomestic;
+    private BigDecimal annualFeeDomestic;
 }
